+++ package/src/components/MessageInput/MessageInput.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { StyleSheet, View } from 'react-native';
+import { Alert, StyleSheet, View } from 'react-native';
 
 import type { UserResponse } from 'stream-chat';
 
 import { useCountdown } from './hooks/useCountdown';
 
+import { ChatContextValue, useChatContext } from '../../contexts';
 import { useAttachmentPickerContext } from '../../contexts/attachmentPickerContext/AttachmentPickerContext';
 import {
   ChannelContextValue,
@@ -78,7 +79,8 @@ const styles = StyleSheet.create({
 
 type MessageInputPropsWithContext<
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
-> = Pick<ChannelContextValue<StreamChatGenerics>, 'disabled' | 'members' | 'watchers'> &
+> = Pick<ChatContextValue<StreamChatGenerics>, 'isOnline'> &
+  Pick<ChannelContextValue<StreamChatGenerics>, 'disabled' | 'members' | 'watchers'> &
   Pick<
     MessageInputContextValue<StreamChatGenerics>,
     | 'additionalTextInputProps'
@@ -159,6 +161,7 @@ const MessageInputWithContext = <
     InputEditingStateHeader,
     InputGiphySearch,
     InputReplyStateHeader,
+    isOnline,
     isValidMessage,
     maxNumberOfFiles,
     members,
@@ -270,15 +273,31 @@ const MessageInputWithContext = <
     }
   }, [imagesForInput]);
 
+  const MEGA_BYTES_TO_BYTES = 1024 * 1024;
+  const MAX_FILE_SIZE_TO_UPLOAD_IN_MB = 100;
+
   const uploadImagesHandler = () => {
-    const imagesToUpload = selectedImages.filter((selectedImage) => {
+    const imageToUpload = selectedImages.find((selectedImage) => {
       const uploadedImage = imageUploads.find(
         (imageUpload) =>
           imageUpload.file.uri === selectedImage.uri || imageUpload.url === selectedImage.uri,
       );
       return !uploadedImage;
     });
-    imagesToUpload.forEach((image) => uploadNewImage(image));
+    // Check if the file size of the image exceeds the threshold of 100MB
+    if (
+      imageToUpload &&
+      Number(imageToUpload.fileSize) / MEGA_BYTES_TO_BYTES > MAX_FILE_SIZE_TO_UPLOAD_IN_MB
+    ) {
+      Alert.alert(
+        `Maximum file size upload limit reached, please upload an image below ${MAX_FILE_SIZE_TO_UPLOAD_IN_MB}MB.`,
+      );
+      setSelectedImages(
+        selectedImages.filter((selectedImage) => selectedImage.uri !== imageToUpload.uri),
+      );
+    } else {
+      if (imageToUpload) uploadNewImage(imageToUpload);
+    }
   };
 
   const removeImagesHandler = () => {
@@ -307,14 +326,27 @@ const MessageInputWithContext = <
   useEffect(() => {
     if (selectedFilesLength > fileUploadsLength) {
       /** User selected a video in bottom sheet attachment picker */
-      const filesToUpload = selectedFiles.filter((selectedFile) => {
+      const fileToUpload = selectedFiles.find((selectedFile) => {
         const uploadedFile = fileUploads.find(
           (fileUpload) =>
             fileUpload.file.uri === selectedFile.uri || fileUpload.url === selectedFile.uri,
         );
         return !uploadedFile;
       });
-      filesToUpload.forEach((file) => uploadNewFile(file));
+      // Check if the file size exceeds the threshold of 100MB
+      if (
+        fileToUpload &&
+        Number(fileToUpload.size) / MEGA_BYTES_TO_BYTES > MAX_FILE_SIZE_TO_UPLOAD_IN_MB
+      ) {
+        Alert.alert(
+          `Maximum file size upload limit reached, please upload a file below ${MAX_FILE_SIZE_TO_UPLOAD_IN_MB}MB.`,
+        );
+        setSelectedFiles(
+          selectedFiles.filter((selectedFile) => selectedFile.uri !== fileToUpload.uri),
+        );
+      } else {
+        if (fileToUpload) uploadNewFile(fileToUpload);
+      }
     } else {
       /** User de-selected a video in bottom sheet attachment picker */
       const filesToRemove = fileUploads.filter(
@@ -526,7 +558,7 @@ const MessageInputWithContext = <
                 ) : null}
                 {fileUploads.length ? <FileUploadPreview /> : null}
                 {giphyActive ? (
-                  <InputGiphySearch />
+                  <InputGiphySearch disabled={!isOnline} />
                 ) : (
                   <View style={[styles.autoCompleteInputContainer, autoCompleteInputContainer]}>
                     <AutoCompleteInput<StreamChatGenerics>
@@ -540,7 +572,11 @@ const MessageInputWithContext = <
                 {cooldownRemainingSeconds ? (
                   <CooldownTimer seconds={cooldownRemainingSeconds} />
                 ) : (
-                  <SendButton disabled={disabled || sending.current || !isValidMessage()} />
+                  <SendButton
+                    disabled={
+                      disabled || sending.current || !isValidMessage() || (giphyActive && !isOnline)
+                    }
+                  />
                 )}
               </View>
             </>
@@ -598,6 +634,7 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
     fileUploads: prevFileUploads,
     giphyActive: prevGiphyActive,
     imageUploads: prevImageUploads,
+    isOnline: prevIsOnline,
     isValidMessage: prevIsValidMessage,
     mentionedUsers: prevMentionedUsers,
     quotedMessage: prevQuotedMessage,
@@ -616,6 +653,7 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
     fileUploads: nextFileUploads,
     giphyActive: nextGiphyActive,
     imageUploads: nextImageUploads,
+    isOnline: nextIsOnline,
     isValidMessage: nextIsValidMessage,
     mentionedUsers: nextMentionedUsers,
     quotedMessage: nextQuotedMessage,
@@ -660,6 +698,9 @@ const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = Default
 
   const showMoreOptionsEqual = prevShowMoreOptions === nextShowMoreOptions;
   if (!showMoreOptionsEqual) return false;
+
+  const isOnlineEqual = prevIsOnline === nextIsOnline;
+  if (!isOnlineEqual) return false;
 
   const isValidMessageEqual = prevIsValidMessage() === nextIsValidMessage();
   if (!isValidMessageEqual) return false;
@@ -719,6 +760,7 @@ export const MessageInput = <
 >(
   props: MessageInputProps<StreamChatGenerics>,
 ) => {
+  const { isOnline } = useChatContext();
   const ownCapabilities = useOwnCapabilitiesContext();
 
   const { disabled = false, members, watchers } = useChannelContext<StreamChatGenerics>();
@@ -809,6 +851,7 @@ export const MessageInput = <
         InputEditingStateHeader,
         InputGiphySearch,
         InputReplyStateHeader,
+        isOnline,
         isValidMessage,
         maxNumberOfFiles,
         members,