+++ examples/SampleApp/CHANGELOG.md
@@ -1,5 +1,66 @@
 # Change Log
 
+## [1.16.0](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.15.1...sampleapp@v1.16.0) (2022-12-23)
+
+
+### Features
+
+* **SampleApp:** Add dev menu item to clear the local DB ([#1874](https://github.com/GetStream/stream-chat-react-native/issues/1874)) ([2d727ac](https://github.com/GetStream/stream-chat-react-native/commit/2d727ac556e49eef00c8da08b8d1450e50684ffb))
+* initial unread count example ([#1868](https://github.com/GetStream/stream-chat-react-native/issues/1868)) ([d283267](https://github.com/GetStream/stream-chat-react-native/commit/d283267db8bd3d577a9c46d02d0e2894b8341915))
+
+
+### Bug Fixes
+
+* always watch channel before passing to channel component ([#1869](https://github.com/GetStream/stream-chat-react-native/issues/1869)) ([4ae9dce](https://github.com/GetStream/stream-chat-react-native/commit/4ae9dceb418c75176277ce285e861549cafcd6b6))
+
+
+### Workspaces
+
+* Following linked packages updated: [stream-chat-react-native]
+
+### [1.15.1](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.15.0...sampleapp@v1.15.1) (2022-12-03)
+
+
+### Workspaces
+
+* Following linked packages updated: [stream-chat-react-native]
+
+## [1.15.0](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.14.4...sampleapp@v1.15.0) (2022-12-01)
+
+
+### Features
+
+* optimistic DB and UI updates for reactions and messages ([#1758](https://github.com/GetStream/stream-chat-react-native/issues/1758)) ([abe2b70](https://github.com/GetStream/stream-chat-react-native/commit/abe2b70d91bcfb689daa2b7bbf6be205c038fe57))
+
+
+### Workspaces
+
+* Following linked packages updated: [stream-chat-react-native]
+
+### [1.14.4](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.14.3...sampleapp@v1.14.4) (2022-11-18)
+
+
+### Bug Fixes
+
+* **sample-app:** xcode 14 build issues ([7900ee7](https://github.com/GetStream/stream-chat-react-native/commit/7900ee720dd65684575a776cba6702d48b1bb3eb))
+
+
+### Workspaces
+
+* Following linked packages updated: [stream-chat-react-native]
+
+### [1.14.3](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.14.2...sampleapp@v1.14.3) (2022-11-15)
+
+
+### Bug Fixes
+
+* **SampleApp:** update references to renamed functions from image gallery context ([#1780](https://github.com/GetStream/stream-chat-react-native/issues/1780)) ([11bbe74](https://github.com/GetStream/stream-chat-react-native/commit/11bbe74741b2a6129f9abbc1e42803e683df1175))
+
+
+### Workspaces
+
+* Following linked packages updated: [stream-chat-react-native]
+
 ### [1.14.2](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v1.14.1...sampleapp@v1.14.2) (2022-10-21)